# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: IMAGINATE e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/track:
    get:
      operationId: trackOrder
      tags: [orders]
      summary: Track an order by order number and phone
      parameters:
        - name: orderNumber
          in: query
          required: true
          schema:
            type: string
        - name: phone
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order tracking info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderTracking"
        "404":
          description: Not found
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete order (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/tracking:
    post:
      operationId: addTrackingEvent
      tags: [orders]
      summary: Add tracking event (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackingEventInput"
      responses:
        "201":
          description: Event added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingEvent"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    OrderItem:
      type: object
      required: [id, productId, productName, price, quantity]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        price:
          type: number
        quantity:
          type: integer
        size:
          type: ["string", "null"]
        color: { type: [string, "null"] }
        variantId: { type: [string, "null"] }
        isPreOrder:
          type: boolean
    Order:
      type: object
      required: [id, status, total, createdAt, items]
      properties:
        id:
          type: integer
        orderNumber:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        customerAddress:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        courierName:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
        delayReason:
          type: ["string", "null"]
        delayedUntil:
          type: ["string", "null"]
        cancelReason:
          type: ["string", "null"]
        refundInitiated:
          type: ["boolean", "null"]
        status:
          type: string
          enum: [pending, confirmed, packed, preparing, shipped, out_for_delivery, delivered, cancelled, delayed, pre_order]
        total:
          type: number
        hasPreOrder:
          type: boolean
        trackingNote:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderInput:
      type: object
      required: [customerName, customerPhone, customerAddress]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        paymentMethod:
          type: string
    OrderUpdate:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerAddress:
          type: string
        paymentMethod:
          type: string
        courierName:
          type: string
        estimatedDelivery:
          type: string
        status:
          type: string
        trackingNote:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, confirmed, packed, preparing, shipped, out_for_delivery, delivered, cancelled, delayed, pre_order]
        delayReason:
          type: string
        delayedUntil:
          type: string
        cancelReason:
          type: string
        refundInitiated:
          type: boolean
    TrackingEvent:
      type: object
      required: [id, orderId, status, createdAt]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        status:
          type: string
        note:
          type: ["string", "null"]
        createdAt:
          type: string
    TrackingEventInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
        note:
          type: string
    OrderTracking:
      type: object
      required: [orderNumber, status, total, createdAt, items]
      properties:
        orderNumber:
          type: string
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        customerAddress:
          type: ["string", "null"]
        status:
          type: string
        courierName:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
        total:
          type: number
        trackingNote:
          type: ["string", "null"]
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        events:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
...