# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * IMAGINATE e-commerce API
 * OpenAPI spec version: 0.2.0
 */
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  /** @nullable */
  orderNumber?: string | null;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  customerAddress?: string | null;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  courierName?: string | null;
  /** @nullable */
  estimatedDelivery?: string | null;
  /** @nullable */
  delayReason?: string | null;
  /** @nullable */
  delayedUntil?: string | null;
  /** @nullable */
  cancelReason?: string | null;
  /** @nullable */
  refundInitiated?: boolean | null;
  status: OrderStatus;
  total: number;
  hasPreOrder?: boolean;
  /** @nullable */
  trackingNote?: string | null;
  createdAt: string;
  items: OrderItem[];
}